Fixed encapsulation design flaw
Graph.addVertex(Vertex<T>* vert) would have allowed the user to add a vertex from a different graph into the graph, meaning all the neighbors wouldn't have been in the graph too creating a huge problem.

That function is now private and the GraphAlgorithms class is now a friend class of Graph so it has access to that method. Need this for DFS and probably others